added return to default position if player goes out of bounds

+++ PA9_Cat_Apocalypse/Character.hpp
@@ -31,11 +31,11 @@ public:
 		hitbox.setPosition(newPosition);
 	}
 
-	int getX() {
+	float getX() {
 		return hitbox.getPosition().x;
 	}
 
-	int getY() {
+	float getY() {
 		return hitbox.getPosition().y;
 	}
